[naga wgsl-in] Move all `to_wgsl` functions into their own module.
This is code motion only, with minor doc fixes. There should be no
changes to the code.
